pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.18",
    "asyncio>=3.4.3",
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "google-oauth2-tool>=0.0.3",
    "gspread>=6.2.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=22.0",
]
